Extend integration test suite, rename test file
